README.md
Bitget referral code BGB101 is your ultimate key to unlocking an exceptional start on one of the world's leading crypto exchanges. This code has been widely verified by countless users and is consistently recognized as the best to receive maximum incentives upon registration, including a substantial promotion up to 6200 USDT and discount up to 50% on trading fees. For new users stepping into the dynamic world of cryptocurrency in 2026, using BGB101 means not just opening an account, but gaining immediate access to a wealth of opportunities and rewards that elevate your initial trading experience on Bitget.